하위 전체 클라이언트 수정
