+++ src/app/search/shorlog/page.tsx
@@ -1,28 +1,10 @@
-'use client';
+import { Suspense } from 'react';
+import ShorlogSearchPageClient from './ShorlogSearchPageClient';
 
-import { useRouter, useSearchParams } from 'next/navigation';
-import { useEffect } from 'react';
-
-export default function ShortlogSearchPage() {
-  const params = useSearchParams();
-  const router = useRouter();
-
-  const keyword = params.get('keyword') || '';
-  const sort = params.get('sort');
-
-  // ⭐ 기본 정렬값 지정: latest
-  useEffect(() => {
-    if (!sort) {
-      router.replace(`/search/shorlog?keyword=${encodeURIComponent(keyword)}&sort=latest`);
-    }
-  }, [sort, keyword]);
-
-  if (!sort) return null; // redirect 처리 중
-
+export default function Page() {
   return (
-    <div>
-      <h1 className="text-lg font-semibold">숏로그 검색 결과</h1>
-      {/* 숏로그 검색 API 호출 로직 */}
-    </div>
+    <Suspense fallback={null}>
+      <ShorlogSearchPageClient />
+    </Suspense>
   );
 }
